+++ Sources/KeyboardKit/Dictation/Dictation.swift
@@ -12,7 +12,7 @@ import Foundation
  This namespace contains dictation-related types, as well as
  disabled service implementations.
  
- The namespace doesn't contain protocols or open classes, or
- types that are meant to be top-level ones.
+ This namespace doesn't contain protocols, open classes, nor
+ types of higher importance.
  */
 public struct Dictation {}