+++ Sources/KeyboardKit/Extensions/String+Sentences.swift
@@ -9,6 +9,24 @@
 import Foundation
 
 public extension String {
+ 
+    /// A list of mutable, western sentence delimiters.
+    static var sentenceDelimiters = ["!", ".", "?"]
+    
+    /// Whether or not this is a western sentence delimiter.
+    var isSentenceDelimiter: Bool {
+        Self.sentenceDelimiters.contains(self)
+    }
+}
+
+public extension Collection where Element == String {
+
+    /// A list of mutable western sentence delimiters.
+    static var sentenceDelimiters: [String] { String.sentenceDelimiters }
+}
+
+
+public extension String {
 
     /**
      Check whether or not the last character within a string