+++ Tests/KeyboardKitTests/Images/Image+KeyboardKitTests.swift
@@ -18,6 +18,10 @@ class Image_KeyboardKitTests: XCTestCase {
 
     func testKeyboardImagesAreDefined() {
         XCTAssertEqual(result(for: .keyboard), Image(systemName: "keyboard"))
+        XCTAssertEqual(result(for: .keyboardArrowLeft), Image(systemName: "arrow.left"))
+        XCTAssertEqual(result(for: .keyboardArrowRight), Image(systemName: "arrow.right"))
+        XCTAssertEqual(result(for: .keyboardArrowUp), Image(systemName: "arrow.up"))
+        XCTAssertEqual(result(for: .keyboardArrowDown), Image(systemName: "arrow.down"))
         XCTAssertEqual(result(for: .keyboardBackspace), Image(systemName: "delete.left"))
         XCTAssertEqual(result(for: .keyboardBackspaceRtl), Image(systemName: "delete.right"))
         XCTAssertEqual(result(for: .keyboardCommand), Image(systemName: "command"))
@@ -28,12 +32,10 @@ class Image_KeyboardKitTests: XCTestCase {
         XCTAssertEqual(result(for: .keyboardImages), Image(systemName: "photo"))
         XCTAssertEqual(result(for: .keyboardDismiss), Image(systemName: "keyboard.chevron.compact.down"))
         XCTAssertEqual(result(for: .keyboardEmojiSymbol), Image(systemName: "face.smiling"))
-        XCTAssertEqual(result(for: .keyboardLeft), Image(systemName: "arrow.left"))
         XCTAssertEqual(result(for: .keyboardNewline), Image(systemName: "arrow.turn.down.left"))
         XCTAssertEqual(result(for: .keyboardNewlineRtl), Image(systemName: "arrow.turn.down.right"))
         XCTAssertEqual(result(for: .keyboardOption), Image(systemName: "option"))
         XCTAssertEqual(result(for: .keyboardRedo), Image(systemName: "arrow.uturn.right"))
-        XCTAssertEqual(result(for: .keyboardRight), Image(systemName: "arrow.right"))
         XCTAssertEqual(result(for: .keyboardSettings), Image(systemName: "gearshape"))
         XCTAssertEqual(result(for: .keyboardShiftCapslocked), Image(systemName: "capslock.fill"))
         XCTAssertEqual(result(for: .keyboardShiftLowercased), Image(systemName: "shift"))